chore: address assorted clippy warnings
Mechanical cleanups to keep clippy -D warnings moving forward:
- Replace manual min/max patterns with clamp
- Simplify map_or usage and redundant casts
- Use idiomatic helpers (abs_diff, sort_by_key, Error::other)
- Minor refactors/type aliases to reduce type complexity

+++ crates/nova-jdwp/src/wire/client.rs
@@ -444,7 +444,7 @@ impl JdwpClient {
         w.write_u32(slots.len() as u32);
         for (slot, signature) in slots {
             w.write_u32(*slot);
-            w.write_u8(signature_to_tag(signature) as u8);
+            w.write_u8(signature_to_tag(signature));
         }
         let payload = self.send_command_raw(16, 1, w.into_vec()).await?;
         let mut r = JdwpReader::new(&payload);
